% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_cs_attainable}
\alias{susie_get_cs_attainable}
\title{Obtain Credible Sets with Attainable Coverage and Entropy Filtering}
\usage{
susie_get_cs_attainable(res, coverage = 0.95, ethres = 20, ...)
}
\arguments{
\item{res}{A susie fit, typically an output from \code{\link{susie}} or one of its variants.}

\item{coverage}{A number between 0 and 1 specifying the desired coverage of each CS. Default is 0.95.}

\item{ethres}{The entropy threshold. CSs with entropy values greater than or equal to \code{log(ethres)}
will be filtered out. Default is 20.}

\item{...}{Additional arguments to be passed to \code{\link{susie_get_cs}}.}
}
\value{
A list with the following elements:
  \describe{
    \item{cs}{A list in which each element is a vector containing the indices of the variables in the CS.}
    \item{coverage}{The coverage of each CS.}
    \item{requested_coverage}{The requested coverage specified by the \code{coverage} argument.}
  }
}
\description{
This function computes credible sets (CSs) from a fitted SuSiE model,
  filtering out CSs that do not meet the specified coverage and entropy thresholds.
  It returns the CSs, their coverage, and the requested coverage.
}
\details{
The function performs the following steps:
  \enumerate{
    \item Computes the attainable coverage by selecting the maximum value in each column of \code{res$alpha}.
    \item Calculates the coverage attained by summing the attainable coverage values for each row.
    \item Calculates the entropy of each row in the attainable coverage matrix using the \code{entropy} function.
    \item Filters the rows based on the conditions: coverage attained > \code{coverage} and entropy < \code{log(ethres)}.
    \item If no rows satisfy the conditions, it returns a list with \code{cs = NULL}, \code{coverage = NULL},
          and \code{requested_coverage = coverage}.
    \item If rows satisfy the conditions, it creates a filtered version of the \code{res} object with only the selected rows.
    \item Calls \code{susie_get_cs} with the filtered \code{res} object and the desired \code{coverage},
          along with any additional arguments passed through \code{...}.
  }
}
\examples{
\dontrun{
# Fit a SuSiE model
s <- susie(X, y, L = 10)

# Obtain credible sets with attainable coverage and entropy filtering
cs_attainable <- susie_get_cs_attainable(s, coverage = 0.9, ethres = 15)
}

}
\seealso{
\code{\link{susie_get_cs}} for obtaining credible sets without attainable coverage and entropy filtering.
}
